one level of refactoring.
